SHIFT & click to delete placed Tiles, major bug fixes
LEFT/RIGHT SHIFTS and click on a tile to delete it - half of its value rounded down will be returned, and it will cascade. Cascading tiles do not give you a return.

ENEMIES RETURN TO SHIPMODE and do new pathfinding if the tile it's on is null.

Polluting tiles continue to pollute without enemy on it.

Bug fixed with subtracting one more count then it should (cascading) - counts of each type of tile are now correct.